Use mm/s in ConvoyFollower

+++ lib/ConvoyFollower/src/DrivingState.h
@@ -102,19 +102,19 @@ public:
     /**
      * Set maximum motor speed.
      *
-     * @param[in] maxSpeed Maximum motor speed.
+     * @param[in] maxSpeed Maximum motor speed in mm/s.
      */
-    void setMaxMotorSpeed(int16_t maxSpeed);
+    void setMaxMotorSpeed(int32_t maxSpeed);
 
     /**
      * Get the calculated motor speed setpoints.
      *
-     * @param[out] leftMotorSpeed The calculated left motor speed.
-     * @param[out] rightMotorSpeed The calculated right motor speed.
+     * @param[out] leftMotorSpeed The calculated left motor speed in mm/s.
+     * @param[out] rightMotorSpeed The calculated right motor speed in mm/s.
      *
      * @returns true if the speed setpoints are valid. Otherwise, false.
      */
-    bool getMotorSpeedSetpoints(int16_t& leftMotorSpeed, int16_t& rightMotorSpeed) const;
+    bool getMotorSpeedSetpoints(int32_t& leftMotorSpeed, int32_t& rightMotorSpeed) const;
 
     /**
      * Set latest vehicle data.
@@ -235,14 +235,14 @@ private:
     /** Flag: State is active. */
     bool m_isActive;
 
-    /** Maximum motor speed. */
-    int16_t m_maxMotorSpeed;
+    /** Maximum motor speed in mm/s. */
+    int32_t m_maxMotorSpeed;
 
-    /** Calculated left motor speed. */
-    int16_t m_leftMotorSpeed;
+    /** Calculated left motor speed in mm/s. */
+    int32_t m_leftMotorSpeed;
 
-    /** Calculated left motor speed. */
-    int16_t m_rightMotorSpeed;
+    /** Calculated left motor speed in mm/s. */
+    int32_t m_rightMotorSpeed;
 
     /** Latest vehicle data. */
     Telemetry m_vehicleData;
